test(sdk_tools): 分文件内的跨被测对象用例归位到各自主体文件
split_narration_segments 的 7 个视频入队准入用例迁入新建
test_enqueue_videos_admission.py（enqueue_videos 全批准入语义，
与 test_enqueue_videos.py 分开以守住 3000 行上限）；4 个
normalize_drama_script / split_reference_video_units /
generate_episode_script 用例迁回 test_text_generation.py 与
test_split_reference_video_units.py。enqueue_narration_audio 的
3 个 enqueue_videos 用例迁入 test_enqueue_videos.py，1 个
generate_storyboards 用例迁入 test_enqueue_storyboards.py。

纯位移：断言正文、pytestmark 作用域与 node-id 后缀均未改，
tests/integration/server 收集清单前后逐名一致，1728 条不变。
